done 7

+++ src/main/java/lk/ijse/Pharmacy/model/StockUpdate.java
@@ -10,5 +10,5 @@ import lombok.NoArgsConstructor;
 
 public class StockUpdate {
     private String id;
-    private Integer qty;
+    private double qty;
 }
